split types to configTypes and sharedTypes
